move map to new page
